test(grdb): extract shared throwing counting conversion service
Apply round-2 review finding: the per-call throwable
`CountingConversionService` and `FailureLog` helpers were
duplicated verbatim between
`GRDBExpenseBreakdownAssembleTests.swift` and
`GRDBCategoryBalancesAssembleTests.swift`. Future changes would
need fixes in two places.

Extract both into
`MoolahTests/Support/ThrowingCountingConversionService.swift`:

- `final class ThrowingCountingConversionService:
  InstrumentConversionService` — renamed from the per-test
  `CountingConversionService` to avoid collision with the
  existing `actor CountingConversionService` in
  `MoolahTests/Support/CountingConversionService.swift` (which
  is a different shape: always-succeeds, rate-keyed).
- `final class FailureLog: Sendable` — `OSAllocatedUnfairLock`-
  backed Int collector for per-row failure-callback fan-out.

Both are `internal` so sibling test files can use them.

Remove the now-unused `import os` from both test files (verified
no other `OSAllocatedUnfairLock` / `Logger` / `os.*` usage there).
