+++ MoolahTests/Backends/GRDB/GRDBExpenseBreakdownAssembleTests.swift
@@ -1,6 +1,5 @@
 import Foundation
 import Testing
-import os
 
 @testable import Moolah
 
@@ -10,72 +9,15 @@ import os
 /// buckets into financial months.
 ///
 /// These tests drive the static helper directly with an injected
-/// throwing conversion service so the per-row error contract required
-/// by `INSTRUMENT_CONVERSION_GUIDE.md` Rule 11 is captured by a unit
-/// test (no GRDB stack needed). A future refactor that collapses the
-/// per-row `do/catch` shape — e.g. wrapping the entire loop in an outer
-/// catch and logging once — will trip the failure-count and rethrow
-/// assertions below.
+/// throwing conversion service (`ThrowingCountingConversionService` from
+/// `MoolahTests/Support/`) so the per-row error contract required by
+/// `INSTRUMENT_CONVERSION_GUIDE.md` Rule 11 is captured by a unit test
+/// (no GRDB stack needed). A future refactor that collapses the per-row
+/// `do/catch` shape — e.g. wrapping the entire loop in an outer catch
+/// and logging once — will trip the failure-count and rethrow assertions
+/// below.
 @Suite("GRDBAnalysisRepository.assembleExpenseBreakdown — Rule 11 contract")
 struct GRDBExpenseBreakdownAssembleTests {
-  /// Counts conversion calls and lets the caller decide whether to
-  /// throw on each one. The closure receives the call index so a
-  /// caller can fail only specific rows (e.g. "throw on row 0 and 2").
-  /// The counter is guarded by `OSAllocatedUnfairLock` (async-safe,
-  /// `Sendable`) so the service can be used from any isolation domain.
-  private final class CountingConversionService: InstrumentConversionService {
-    private let counter = OSAllocatedUnfairLock(initialState: 0)
-    private let outcome: @Sendable (Int) -> Result<Decimal, any Error>
-
-    init(outcome: @escaping @Sendable (Int) -> Result<Decimal, any Error>) {
-      self.outcome = outcome
-    }
-
-    var calls: Int { counter.withLock { $0 } }
-
-    func convert(
-      _ quantity: Decimal, from: Instrument, to: Instrument, on date: Date
-    ) async throws -> Decimal {
-      let index = counter.withLock { count -> Int in
-        let current = count
-        count += 1
-        return current
-      }
-      switch outcome(index) {
-      case .success(let value):
-        return value
-      case .failure(let error):
-        throw error
-      }
-    }
-
-    func convertAmount(
-      _ amount: InstrumentAmount, to instrument: Instrument, on date: Date
-    ) async throws -> InstrumentAmount {
-      let value = try await convert(
-        amount.quantity, from: amount.instrument, to: instrument, on: date)
-      return InstrumentAmount(quantity: value, instrument: instrument)
-    }
-  }
-
-  /// Async-safe collector for the per-row failure-callback fan-out
-  /// observed by each test. Backed by `OSAllocatedUnfairLock` so the
-  /// `@Sendable` closure passed to
-  /// `ExpenseBreakdownHandlers.handleConversionFailure` can append
-  /// from whichever isolation domain the helper resumes on without a
-  /// data-race waiver.
-  private final class FailureLog: Sendable {
-    private let entries = OSAllocatedUnfairLock<[Int]>(initialState: [])
-
-    func append(_ value: Int) {
-      entries.withLock { $0.append(value) }
-    }
-
-    func snapshot() -> [Int] {
-      entries.withLock { $0 }
-    }
-  }
-
   /// Build a three-row aggregation of USD legs over consecutive days
   /// against the default test instrument as the profile target. Three
   /// rows is enough to observe both per-row callback fan-out and
@@ -103,7 +45,7 @@ struct GRDBExpenseBreakdownAssembleTests {
   @Test("handleConversionFailure invoked once per failing row before rethrow")
   func handleConversionFailureFiresPerRow() async throws {
     let aggregation = makeAggregation()
-    let conversionService = CountingConversionService { index in
+    let conversionService = ThrowingCountingConversionService { index in
       .failure(CallbackError(index: index))
     }
     let failures = FailureLog()
@@ -135,7 +77,7 @@ struct GRDBExpenseBreakdownAssembleTests {
   @Test("loop processes all rows even when the first row fails")
   func loopContinuesAfterFirstFailure() async throws {
     let aggregation = makeAggregation()
-    let conversionService = CountingConversionService { index in
+    let conversionService = ThrowingCountingConversionService { index in
       index == 0 ? .failure(CallbackError(index: index)) : .success(0)
     }
     let visited = FailureLog()
@@ -167,7 +109,7 @@ struct GRDBExpenseBreakdownAssembleTests {
   @Test("CancellationError rethrown immediately without invoking handleConversionFailure")
   func cancellationErrorIsNotFoldedIntoConversionFailureLog() async throws {
     let aggregation = makeAggregation()
-    let conversionService = CountingConversionService { index in
+    let conversionService = ThrowingCountingConversionService { index in
       index == 0 ? .failure(CancellationError()) : .success(0)
     }
     let visited = FailureLog()
